refactor: extract buttons to individual file
