+++ src/modules/home/components/book.js
@@ -8,7 +8,11 @@ import { FormButton, FormOutlineButton } from '../../shared/components/buttons'
 
 import { styles } from './styles/home.styles'
 
-export const Book = ({ onBuyPressed, onRentPressed, book: { title, author, edition, imageUri } }) => (
+export const Book = ({
+  onBuyPressed,
+  onRentPressed,
+  book: { title, author, edition, imageUri }
+}) => (
   <Card style={styles.bookCard}>
     <Image style={styles.bookImage} source={{ uri: imageUri }} />
     <View style={styles.bookInfoWrap}>
@@ -16,8 +20,16 @@ export const Book = ({ onBuyPressed, onRentPressed, book: { title, author, editi
       <Text style={styles.secondaryInput}>{author}</Text>
       <Text style={styles.secondaryInput}>{edition}</Text>
       <View style={styles.bookButtonsGroup}>
-        <FormButton style={styles.buyButton} title={'Buy'} onPress={onBuyPressed} />
-        <FormOutlineButton style={styles.rentButton} title={'Rent'} onPress={onRentPressed} />
+        <FormButton
+          style={styles.buyButton}
+          title={'Buy'}
+          onPress={onBuyPressed}
+        />
+        <FormOutlineButton
+          style={styles.rentButton}
+          title={'Rent'}
+          onPress={onRentPressed}
+        />
       </View>
     </View>
   </Card>
