--- src/modules/shared/components/buttons.js
@@ -1,68 +1,0 @@
-import React from 'react'
-import { func } from 'prop-types'
-import { Text, StyleSheet, TouchableHighlight } from 'react-native'
-
-import { styles } from './styles/buttons.styles'
-import { Colors } from '../../../constants'
-
-// TODO: Refactor this
-
-const propTypes = {
-  onPress: func.isRequired
-}
-
-export const FormButton = props => {
-  const { primaryButton, darkenOutline, whiteText } = styles
-
-  const style = StyleSheet.flatten([primaryButton, darkenOutline, props.style])
-  return (
-    <TouchableHighlight
-      onPress={props.onPress}
-      style={style}
-      underlayColor={Colors.primary700}
-    >
-      <Text style={whiteText}>{props.title}</Text>
-    </TouchableHighlight>
-  )
-}
-
-FormButton.propTypes = propTypes
-
-export const FormOutlineButton = props => {
-  const { transparentButton, primaryOutline, primaryText } = styles
-
-  const style = StyleSheet.flatten([
-    transparentButton,
-    primaryOutline,
-    props.style
-  ])
-  return (
-    <TouchableHighlight
-      onPress={props.onPress}
-      style={style}
-      underlayColor={Colors.primary700}
-    >
-      <Text style={primaryText}>{props.title}</Text>
-    </TouchableHighlight>
-  )
-}
-
-FormOutlineButton.propTypes = propTypes
-
-export const CardFooterButton = props => {
-  const { bottomRadius, primaryButton, whiteText, secondaryButton } = styles
-
-  const buttonColor = props.secondary ? secondaryButton : primaryButton
-  const style = StyleSheet.flatten([bottomRadius, buttonColor, props.style])
-  return (
-    <TouchableHighlight
-      onPress={props.onPress}
-      style={style}
-      underlayColor={Colors.primary700}
-    >
-      <Text style={whiteText}>{props.title}</Text>
-    </TouchableHighlight>
-  )
-}
-
-CardFooterButton.propTypes = propTypes
